+++ changelog.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.2.3] - 2024-04-10
+
+### Changed
+
+- parser refactor
+- update deps
+
+### Added
+
+- lodash [CVE-2021-23337](https://github.com/lodash/lodash/issues/5808)
+
+### Removed
+
+- lodash.template [CVE-2021-23337](https://github.com/lodash/lodash/issues/5808)
+
 ## [5.2.2] - 2024-03-26
 
 ### Changed